+++ apps/web/src/components/lien/forms/add-medical-lien/medical-upload-document.tsx
@@ -18,6 +18,7 @@ import { useLienStore } from "@/stores/lien-store";
 import { ApiError } from "@/lib/api-client";
 import { ConfirmDialog } from "../../modal";
 import { documentsService } from "@/lib/documents";
+import { dateConverter } from "@/lib/cases/cases.mapper";
 
 export interface UploadDocumentsProps {
   caseId?: string;
@@ -26,7 +27,13 @@ export interface UploadDocumentsProps {
   onFormValid?: (valid: boolean, data?: any) => void;
   openAddFundingCompanyModal?: () => void;
   onUploaded?: (valid: boolean, data?: any) => void;
+  mode?: "add" | "edit";
 }
+
+type tempFileType = {
+  file: File;
+  documentTypeId: string;
+};
 
 const INITIAL_FORM = {
   documentType: "",
@@ -98,8 +105,10 @@ export default function UploadDocuments(props: UploadDocumentsProps) {
       label: d.name,
     };
   });
-  const [documents, setDocuments] = useState<any[]>(data);
+  const [documents, setDocuments] = useState<any[]>([]);
   const [files, setFiles] = useState<File[] | null>(null);
+  const [temporaryFiles, setTemporaryFiles] = useState<tempFileType[]>([]);
+
   const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
 
   useEffect(() => {}, [data, isSubmitting]);
@@ -136,6 +145,18 @@ export default function UploadDocuments(props: UploadDocumentsProps) {
     showConfirmAction({ isOpen: true, id: fileId });
   }
   const deleteFile = useCallback(async () => {
+    if (props.mode == "add") {
+      setDocuments((current) =>
+        current.filter((row) => row.id !== confirmAction.id),
+      );
+      addToast({
+        type: "success",
+        title: `Deleted`,
+      });
+      showConfirmAction({ id: "", isOpen: false });
+
+      return;
+    }
     try {
       await casesService.deleteLiensDocument(confirmAction.id);
       fetchDocument();
@@ -148,9 +169,37 @@ export default function UploadDocuments(props: UploadDocumentsProps) {
         });
       }
     }
+    showConfirmAction({ id: "", isOpen: false });
   }, [confirmAction]);
 
+  const setFilesData = (files: File[] | null) => {
+    const tempData =
+      files?.map((f: any) => {
+        return {
+          file: f,
+          documentTypeId: form.documentType,
+        };
+      }) ?? [];
+    setTemporaryFiles((prev) => [...prev, ...tempData]);
+    setFiles(files);
+  };
+
   const uploadLiensDocuments = useCallback(async () => {
+    if (props.mode == "add") {
+      const filesData =
+        files?.map((d: any, index: number) => ({
+          name: d.name,
+          type: form.documentType,
+          id: documents.length,
+          url: d.url,
+        })) ?? [];
+      setDocuments((prev) => [...prev, ...filesData]);
+
+      dropzoneRef?.current?.reset();
+      props.onFormValid?.(true, temporaryFiles);
+      setForm(initialForm);
+      return;
+    }
     setIsSubmitting(true);
 
     if (!files) return;
@@ -174,6 +223,7 @@ export default function UploadDocuments(props: UploadDocumentsProps) {
           setForm(initialForm);
           props?.onUploaded?.(true, "");
           props.onFormValid?.(true, "");
+          fetchDocument();
           setIsSubmitting(false);
         });
       } catch (err) {
@@ -194,13 +244,14 @@ export default function UploadDocuments(props: UploadDocumentsProps) {
         }
       }
     }, 0);
-  }, [isSubmitting, files]);
+  }, [isSubmitting, files, documents]);
 
   const fetchDocument = async () => {
     const docs = await casesService.loadLiensDocuments(lienId ?? "");
     setDocuments(
       docs.data.map((d: any) => {
         return {
+          ...d,
           name: d.filename,
           type: d.typeId,
           id: d.id,
@@ -212,22 +263,23 @@ export default function UploadDocuments(props: UploadDocumentsProps) {
   const getDocumentNameById = (id: string) => {
     return documentTypes?.find((t) => t.key == id)?.label;
   };
+
   useEffect(() => {
-    if (data?.length > 0) {
-      setDocuments(
-        data.map((d: any) => {
-          return {
-            name: d.filename,
-            type: d.typeId,
-            id: d.id,
-            url: d.url,
-          };
-        }),
-      );
-    } else {
-      fetchDocument();
+    if (props.mode == "edit") {
+      if (data.hasInitialValue) {
+        setDocuments(
+          data.files.map((d: any) => {
+            return {
+              name: d.filename,
+              type: d.typeId,
+              id: d.id,
+              url: d.url,
+            };
+          }),
+        );
+      }
     }
-  }, [data, onUploaded]);
+  }, []);
 
   return (
     <div className="container-fluid">
@@ -254,8 +306,7 @@ export default function UploadDocuments(props: UploadDocumentsProps) {
                 Array.isArray(e) &&
                 e.length > 0 &&
                 e.every((item) => item instanceof File);
-
-              setFiles(finalizedFiles ? e : null);
+              setFilesData(finalizedFiles ? e : null);
             }}
           />
           <button
